pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["vikyw89 <vikyw89@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
together = "^1.2.0"
python-dotenv = "^1.0.1"
tenacity = "^8.3.0"


ruff = "^0.4.8"
instructor = "^1.3.2"
litellm = "^1.40.4"
pytest = "^8.2.2"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
start = "scripts.start:run"
lint = "scripts.lint:run"
